#1067 Double -> BigDecimal
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13176 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/opentaps-common/src/common/org/opentaps/common/order/OrderServices.java
@@ -1713,7 +1713,7 @@ public final class OrderServices {
                     shippingTotal = BigDecimal.ZERO;
                     Debug.log("No valid order items found - " + shippingTotal, MODULE);
                 } else {
-                    shippingTotal = UtilValidate.isEmpty(shippingEstMap.get("shippingTotal")) ? BigDecimal.ZERO : BigDecimal.valueOf(((Double) shippingEstMap.get("shippingTotal")).doubleValue());
+                    shippingTotal = UtilValidate.isEmpty(shippingEstMap.get("shippingTotal")) ? BigDecimal.ZERO : (BigDecimal) shippingEstMap.get("shippingTotal");
                     shippingTotal = shippingTotal.setScale(decimals, rounding);
                 }
                 if (Debug.infoOn()) {
